[Perf] Lazy-load framer motion components
